pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "support-bot"
version = "0.1.0"
description = "Telegram bot for automated user support with AI"
authors = ["Broxus <info@broxus.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = ">=22.1,<23.0"
openai = ">=1.82.1,<2.0.0"
pyyaml = ">=6.0.2,<7.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
faiss-cpu = ">=1.11.0,<2.0.0"
openai-agents = ">=0.0.16,<0.0.17"
langchain-community = "^0.3.24"
langchain-openai = "^0.3.21"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.1.1"
ruff = "^0.11.12"
pyyaml = "^6.0.2"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "ERA", # eradicate
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
